switch to Ffuenf_DeleteOrder integrate test-suite
