+++ .gitignore
@@ -1,0 +1,2 @@
+.vagrant
+travis_key_chunk.sh
