+++ .travis.yml
@@ -1,40 +1,21 @@
 language: php
-
-php:
-  - 5.3
-  - 5.4
-  - 5.5
-
 sudo: false
-
+php:
+- 5.3
+- 5.4
+- 5.5
+- 5.6
+- hhvm
+matrix:
+  allow_failures:
+  - php: 5.6
+  - php: hhvm
 env:
-  - MAGENTO_VERSION="1.5.0.1"
-  - MAGENTO_VERSION="1.6.0.0"
-  - MAGENTO_VERSION="1.6.1.0"
-  - MAGENTO_VERSION="1.6.2.0"
-  - MAGENTO_VERSION="1.7.0.0"
-  - MAGENTO_VERSION="1.7.0.1"
-  - MAGENTO_VERSION="1.7.0.2"
-  - MAGENTO_VERSION="1.8.0.0"
-  - MAGENTO_VERSION="1.8.1.0"
-  - MAGENTO_VERSION="1.9.0.0"
-  - MAGENTO_VERSION="1.9.0.1"
-  - MAGENTO_VERSION="1.9.1.0"
-
-before_script:
-  - mkdir magento/
-  - git clone https://github.com/netz98/n98-magerun ./n98-magerun/
-  - git clone git://github.com/EcomDev/MageCI.git ./mageci/
-  - ./mageci/bin/mage-ci install magento $MAGENTO_VERSION magento_test -c -t -r http://mage-ci.ecomdev.org
-  - composer update
-  - git clone git://github.com/EcomDev/EcomDev_PHPUnit.git -b dev ./phpunit/
-  - CURR_DIR=$(pwd)
-  - wget https://raw.githubusercontent.com/colinmollenhour/modman/master/modman -O ./mageci/bin/modman
-  - chmod +x ./mageci/bin/modman
-  - ./mageci/bin/mage-ci install-module magento $CURR_DIR/phpunit
-  - ./mageci/bin/mage-ci install-module magento $CURR_DIR
-  - composer install --dev
-  - cd magento
-
+  matrix:
+  - MAGENTO_VERSION=magento-ce-1.9.1.1
+  - MAGENTO_VERSION=magento-ce-1.8.1.0
+  - MAGENTO_VERSION=magento-ce-1.7.0.2
+  - MAGENTO_VERSION=magento-ce-1.6.2.0
 script:
-  - ../vendor/bin/phpunit --group EM_DeleteOrder
+- curl -sSL https://raw.githubusercontent.com/ffuenf/MageTestStand/master/setup.sh
+  | bash
