+++ CHANGELOG.md
@@ -1,0 +1,8 @@
+# CHANGELOG for Ffuenf_DeleteOrder
+
+This file is used to list changes made in each version of Ffuenf_DeleteOrder.
+
+## 2.0.0 (May 20, 2015)
+
+* switch to Ffuenf_DeleteOrder
+* integrate test-suite
