--- app/code/community/EM/DeleteOrder/Test/Helper/Data/fixtures/testIsExtensionActive.yaml
@@ -1,2 +1,0 @@
-config:
-  default/system/em_deleteorder/enable: 1
