--- app/code/community/EM/DeleteOrder/etc/system.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0"?>
-<config>
-  <sections>
-    <system>
-      <groups>
-        <em_deleteorder translate="label" module="em_deleteorder">
-          <label>EM_DeleteOrder</label>
-          <frontend_type>text</frontend_type>
-          <sort_order>900</sort_order>
-          <show_in_default>1</show_in_default>
-          <show_in_website>0</show_in_website>
-          <show_in_store>0</show_in_store>
-          <fields>
-            <enable translate="label,comment">
-              <label>Enable EM_DeleteOrder</label>
-              <frontend_type>select</frontend_type>
-              <source_model>adminhtml/system_config_source_yesno</source_model>
-              <sort_order>1</sort_order>
-              <show_in_default>1</show_in_default>
-              <show_in_website>0</show_in_website>
-              <show_in_store>0</show_in_store>
-              <comment><![CDATA[Enable or disable extension.]]></comment>
-            </enable>
-          </fields>
-        </em_deleteorder>
-      </groups>
-    </system>
-  </sections>
-</config>
