+++ app/code/local/Ffuenf/DeleteOrder/Helper/Data.php
@@ -1,0 +1,42 @@
+<?php
+/**
+ * Ffuenf_DeleteOrder extension
+ * 
+ * NOTICE OF LICENSE
+ * 
+ * This source file is subject to the MIT License
+ * that is bundled with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/mit-license.php
+ * 
+ * @category   Ffuenf
+ * @package    Ffuenf_DeleteOrder
+ * @author     Achim Rosenhagen <a.rosenhagen@ffuenf.de>
+ * @copyright  Copyright (c) 2015 ffuenf (http://www.ffuenf.de)
+ * @license    http://opensource.org/licenses/mit-license.php MIT License
+*/
+
+class Ffuenf_DeleteOrder_Helper_Data extends Ffuenf_Pagespeed_Helper_Core
+{
+    /**
+     * Path for the config for extension active status
+     */
+    const CONFIG_EXTENSION_ACTIVE = 'deleteorder/general/enabled';
+
+    /**
+     * Variable for if the extension is active
+     *
+     * @var bool
+     */
+    protected $bExtensionActive;
+
+    /**
+     * Check to see if the extension is active
+     *
+     * @return bool
+     */
+    public function isExtensionActive()
+    {
+        return $this->getStoreFlag(self::CONFIG_EXTENSION_ACTIVE, 'bExtensionActive');
+    }
+}
