+++ app/code/local/Ffuenf/DeleteOrder/Test/Helper/Data/fixtures/shared.yaml
@@ -1,0 +1,2 @@
+config:
+  default/deleteorder/general/enabled: 1
