+++ app/code/local/Ffuenf/DeleteOrder/etc/config.xml
@@ -1,0 +1,60 @@
+<?xml version="1.0"?>
+<config>
+    <modules>
+        <Ffuenf_DeleteOrder>
+            <version>2.0.0</version>
+        </Ffuenf_DeleteOrder>
+    </modules>
+    <global>
+        <helpers>
+            <ffuenf_deleteorder>
+                <class>Ffuenf_DeleteOrder_Helper</class>
+            </ffuenf_deleteorder>
+        </helpers>
+        <rewrite>
+            <admin_adminhtml_sales_order>
+                <from><![CDATA[#/admin/sales_order/#]]></from>
+                <to>/deleteorder/adminhtml_sales_order/</to>
+            </admin_adminhtml_sales_order>
+        </rewrite>
+        <blocks>
+            <adminhtml>
+                <rewrite>
+                    <sales_order_grid>Ffuenf_DeleteOrder_Block_Adminhtml_Sales_Order_Grid</sales_order_grid>
+                </rewrite>
+            </adminhtml>
+        </blocks>
+    </global>
+    <admin>
+        <routers>
+            <deleteorder>
+                <use>admin</use>
+                <args>
+                    <module>Ffuenf_DeleteOrder</module>
+                    <frontName>deleteorder</frontName>
+                </args>
+            </deleteorder>
+            <adminhtml>
+                <args>
+                    <modules>
+                        <Ffuenf_DeleteOrder_Adminhtml before="Mage_Adminhtml">Ffuenf_DeleteOrder_Adminhtml</Ffuenf_DeleteOrder_Adminhtml>
+                    </modules>
+                </args>
+            </adminhtml>
+        </routers>
+    </admin>
+    <default>
+        <system>
+            <deleteorder>
+                <enabled>0</enabled>
+            </deleteorder>
+        </system>
+    </default>
+    <phpunit>
+        <suite>
+            <modules>
+                <Ffuenf_DeleteOrder/>
+            </modules>
+        </suite>
+    </phpunit>
+</config>
