+++ app/code/local/Ffuenf/DeleteOrder/etc/system.xml
@@ -1,0 +1,55 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Ffuenf_DeleteOrder extension
+ * 
+ * NOTICE OF LICENSE
+ * 
+ * This source file is subject to the MIT License
+ * that is bundled with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/mit-license.php
+ * 
+ * @category   Ffuenf
+ * @package    Ffuenf_DeleteOrder
+ * @author     Achim Rosenhagen <a.rosenhagen@ffuenf.de>
+ * @copyright  Copyright (c) 2015 ffuenf (http://www.ffuenf.de)
+ * @license    http://opensource.org/licenses/mit-license.php MIT License
+*/
+-->
+<config>
+    <tabs>
+        <ffuenf>
+            <label>ffuenf</label>
+            <sort_order>508</sort_order>
+        </ffuenf>
+    </tabs>
+    <sections>
+        <deleteorder translate="label" module="ffuenf_deleteorder">
+            <tab>ffuenf</tab>
+            <label>Pagespeed</label>
+            <sort_order>10</sort_order>
+            <show_in_default>1</show_in_default>
+            <show_in_website>1</show_in_website>
+            <show_in_store>1</show_in_store>
+            <groups>
+                <general translate="label" module="ffuenf_deleteorder">
+                    <label>General Configuration</label>
+                    <sort_order>1000</sort_order>
+                    <show_in_default>1</show_in_default>
+                    <show_in_website>1</show_in_website>
+                    <show_in_store>1</show_in_store>
+                    <fields>
+                        <enabled translate="label">
+                            <label>Enable</label>
+                            <frontend_type>select</frontend_type>
+                            <source_model>adminhtml/system_config_source_yesno</source_model>
+                            <sort_order>10</sort_order>
+                            <show_in_default>1</show_in_default>
+                        </enabled>
+                    </fields>
+                </general>
+            </groups>
+        </deleteorder>
+    </sections>
+</config>
