--- app/etc/config.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0"?>
-<config>
-  <modules>
-    <EM_DeleteOrder>
-      <version>1.0.0</version>
-    </EM_DeleteOrder>
-  </modules>
-  <global>
-    <rewrite>
-      <em_emadmin_adminhtml_sales_order>
-        <from><![CDATA[#/admin/sales_order/#]]></from>
-        <to>/emadmin/adminhtml_sales_order/</to>
-      </em_emadmin_adminhtml_sales_order>
-    </rewrite>
-    <blocks>
-      <adminhtml>
-        <rewrite>
-          <sales_order_grid>EM_DeleteOrder_Block_Adminhtml_Sales_Order_Grid</sales_order_grid>
-        </rewrite>
-      </adminhtml>
-    </blocks>
-  </global>
-  <admin>
-    <routers>
-      <em_deleteorder>
-        <use>admin</use>
-        <args>
-          <module>EM_DeleteOrder</module>
-          <frontName>emadmin</frontName>
-        </args>
-      </em_deleteorder>
-      <adminhtml>
-        <args>
-          <modules>
-            <EM_DeleteOrder_Adminhtml before="Mage_Adminhtml">EM_DeleteOrder_Adminhtml</EM_DeleteOrder_Adminhtml>
-          </modules>
-        </args>
-      </adminhtml>
-    </routers>
-  </admin>
-  <phpunit>
-    <suite>
-      <modules>
-        <EM_DeleteOrder/>
-      </modules>
-    </suite>
-  </phpunit>
-</config>
