+++ app/etc/modules/Ffuenf_DeleteOrder.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Ffuenf_DeleteOrder extension
+ * 
+ * NOTICE OF LICENSE
+ * 
+ * This source file is subject to the MIT License
+ * that is bundled with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/mit-license.php
+ * 
+ * @category   Ffuenf
+ * @package    Ffuenf_DeleteOrder
+ * @author     Achim Rosenhagen <a.rosenhagen@ffuenf.de>
+ * @copyright  Copyright (c) 2015 ffuenf (http://www.ffuenf.de)
+ * @license    http://opensource.org/licenses/mit-license.php MIT License
+*/
+-->
+<config>
+    <modules>
+        <Ffuenf_DeleteOrder>
+            <active>true</active>
+            <codePool>local</codePool>
+        </Ffuenf_DeleteOrder>
+    </modules>
+</config>
