--- package.xml
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<package packagerversion="1.7.1" version="2.0" xmlns="http://pear.php.net/dtd/package-2.0" xmlns:tasks="http://pear.php.net/dtd/tasks-1.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://pear.php.net/dtd/tasks-1.0
-    http://pear.php.net/dtd/tasks-1.0.xsd
-    http://pear.php.net/dtd/package-2.0
-    http://pear.php.net/dtd/package-2.0.xsd">
- <name>EM_DeleteOrder</name>
- <channel>connect.magentocommerce.com/community</channel>
- <summary>Enhance Magento feature with ability to delete orders</summary>
- <description>Enhance Magento feature with ability to delete orders</description>
- <lead>
-  <name>EMThemes</name>
-  <user>tvlgiao</user>
-  <email>tvlgiao@gmail.com</email>
-  <active>yes</active>
- </lead>
- <date>2011-03-07</date>
- <time>12:51:57</time>
- <version>
-  <release>1.0.5</release>
-  <api>1.0</api>
- </version>
- <stability>
-  <release>stable</release>
-  <api>stable</api>
- </stability>
- <license>GNU/GPL</license>
- <notes>Version 1.0.5:
-- Support delete non-numeric order number.
-
-Version 1.0.4:
-- Refactor directory structure.
-- Fix bug not work if tables have prefix.
-- Fix bug not work if backend url is not admin/</notes>
- <contents>
-  <dir name="/">
-   <dir name="EM">
-    <dir name="DeleteOrder">
-     <dir name="Block">
-      <dir name="Adminhtml">
-       <dir name="Sales">
-        <dir name="Order">
-         <file md5sum="0a9a2559b0d0df86061482451927236a" name="Grid.php" role="magecommunity" />
-        </dir> <!-- /EM/DeleteOrder/Block/Adminhtml/Sales/Order -->
-       </dir> <!-- /EM/DeleteOrder/Block/Adminhtml/Sales -->
-      </dir> <!-- /EM/DeleteOrder/Block/Adminhtml -->
-     </dir> <!-- /EM/DeleteOrder/Block -->
-     <dir name="controllers">
-      <dir name="Adminhtml">
-       <dir name="Sales">
-        <file md5sum="0039cf1dfcaf7eab0a1daf839ae5012e" name="OrderController.php" role="magecommunity" />
-       </dir> <!-- /EM/DeleteOrder/controllers/Adminhtml/Sales -->
-      </dir> <!-- /EM/DeleteOrder/controllers/Adminhtml -->
-     </dir> <!-- /EM/DeleteOrder/controllers -->
-     <dir name="etc">
-      <file md5sum="78b876e905249f6a47e8ce1500b8a5a4" name="config.xml" role="magecommunity" />
-     </dir> <!-- /EM/DeleteOrder/etc -->
-    </dir> <!-- /EM/DeleteOrder -->
-   </dir> <!-- /EM -->
-   <dir name="modules">
-    <file md5sum="9e363a45860cf072385c1a0d24b5bc85" name="EM_DeleteOrder.xml" role="mageetc" />
-   </dir> <!-- /modules -->
-  </dir> <!-- / -->
- </contents>
- <dependencies>
-  <required>
-   <php>
-    <min>5.2</min>
-    <max>6.0</max>
-   </php>
-   <pearinstaller>
-    <min>1.6.2</min>
-   </pearinstaller>
-  </required>
- </dependencies>
- <phprelease />
-</package>
